conversation_app/requirements.txt
# Framework
fastapi==0.122.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
slowapi==0.1.9

# Auth
bcrypt==5.0.0
cryptography==44.0.3
python-jose[cryptography]==3.5.0

# LangChain
langchain==1.2.9
langchain-community==0.4.1
langchain-huggingface==1.2.0
huggingface-hub==0.36.0
faiss-cpu==1.13.2

# Cache & Utils
diskcache==5.6.3
pdfplumber==0.11.9
easyocr==1.7.2
numpy==2.4.2
aiohttp==3.13.2
aiofiles
nest_asyncio
joblib
python-multipart

# LLM (CPU only)
llama-cpp-python==0.3.16
